+++ hippynn/custom_kernels/env_numba.py
@@ -38,6 +38,8 @@ class WrappedEnvsum(NumbaCompatibleTensorFunction):
     def launch_bounds(self, sense_shape, fs, pfs, pss, atom1_ids_shape, *other_shapes):
         n_pairs, n_nu = sense_shape
         n_atom, n_feat = fs
+        if n_feat > 512:
+            raise ValueError(f"Numba GPU custom kernels are not compatible with feature sizes greater than 512 (got {n_feat})")
         (n_atoms_interacting,) = atom1_ids_shape
         TPB_MAX = 512
         TPB_X = n_feat
@@ -50,9 +52,7 @@ class WrappedEnvsum(NumbaCompatibleTensorFunction):
 
     @staticmethod
     def make_kernel(KERNEL_DTYPE):
-        sig = "void({DTYPE}[:,:,],{DTYPE}[:,:],int64[:],int64[:],int64[:],int64[:],{DTYPE}[:,:,:])".format(
-            DTYPE=KERNEL_DTYPE
-        )
+        sig = "void({DTYPE}[:,:,],{DTYPE}[:,:],int64[:],int64[:],int64[:],int64[:],{DTYPE}[:,:,:])".format(DTYPE=KERNEL_DTYPE)
 
         @numba.cuda.jit(
             sig,
@@ -121,11 +121,14 @@ class WrappedSensesum(NumbaCompatibleTensorFunction):
     def launch_bounds(self, env_shape, feat_shape, pfirst_shape, psecond_shape):
         (n_pairs,) = pfirst_shape
         n_atoms, n_nu, n_feat = env_shape
+        if n_nu > 512:
+            raise ValueError(f"Numba GPU custom kernels are not compatible with sensitivity sizes greater than 512 (got {n_nu})")
+
         TPB_MAX = 512
         TPB_Y = n_nu
-        TPB_X = TPB_MAX//TPB_Y
-        TPB = (TPB_X,TPB_Y)
-        BPG = (n_pairs + TPB_X -1 )//TPB_X
+        TPB_X = TPB_MAX // TPB_Y
+        TPB = (TPB_X, TPB_Y)
+        BPG = (n_pairs + TPB_X - 1) // TPB_X
         return BPG, TPB
 
     @staticmethod
@@ -193,6 +196,8 @@ class WrappedFeatsum(NumbaCompatibleTensorFunction):
         n_pairs, n_nu = sense_shape
         n_atom, n_nu, n_feat = env_shape
         TPB_max = 512
+        if n_feat > 512:
+            raise ValueError(f"Numba GPU custom kernels are not compatible with feature sizes greater than 512 (got {n_feat})")
         if n_feat > 32:
             TPB_x = ((n_feat + 31) // 32) * 32
             TPB_y = TPB_max // TPB_x
@@ -209,9 +214,7 @@ class WrappedFeatsum(NumbaCompatibleTensorFunction):
 
     @staticmethod
     def make_kernel(KERNEL_DTYPE):
-        sig = "void({DTYPE}[:,:,:],{DTYPE}[:,:],int64[:],int64[:],int64[:],int64[:],{DTYPE}[:,:])".format(
-            DTYPE=KERNEL_DTYPE
-        )
+        sig = "void({DTYPE}[:,:,:],{DTYPE}[:,:],int64[:],int64[:],int64[:],int64[:],{DTYPE}[:,:])".format(DTYPE=KERNEL_DTYPE)
 
         @numba.cuda.jit(
             sig,